requirements.txt
flask
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-community
chromadb
markdown